Improve code substituting \o, \O etc.
These broke TeX macros like `\overline`. To fix that, check for a word
boundary after the letter. But now we have to be careful to perform these
substitutions first and together. Otherwise, if we have e.g. `\o\l` and first
replace the `\l` by `ł`, then the result would `\oł` which has no word boundary
after `o` anymore, and we'd not substitute the `\o` anymore.